update ID Number type

+++ Models/Customer.cs
@@ -23,7 +23,7 @@ namespace TweakBank.Models
         public string? Country { get; set; }
         public string? Phone { get; set; }
 
-        public int IdNumber { get; set; }
+        public long IdNumber { get; set; }
 
     }
 }
